+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/typescript/Typescript.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2023 DiffPlug
+ * Copyright 2016-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -48,10 +48,5 @@ public class Typescript extends FormatterFactory {
 
 	public void addBiome(BiomeTs biome) {
 		addStepFactory(biome);
-	}
-
-	@Deprecated
-	public void addRome(RomeTs rome) {
-		addStepFactory(rome);
 	}
 }